+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2021-11-02
+### Added
+- `devatherock/nexmo-oas-renderer` to visualize api specs from a git repo
+
 ## 2020-11-14
 ### Added
 - `shasum` to `devatherock/graalvm`